% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation_metrics.R
\name{citation_metrics}
\alias{citation_metrics}
\title{Fetch CrossRef bibliometrics for PubMed records.}
\usage{
citation_metrics(df)
}
\arguments{
\item{df}{The dataframe object returned by \code{\link{get_records}}. Alternatively, any dataframe that includes a column named \emph{pmid}, which contains the PMIDs of each PubMed record bibliometrics are required for.}
}
\value{
Returns a dataframe with all the columns of the input dataframe, plus:
\describe{
\item{crossref_citations}{CrossRef citation count \emph{(class: integer)}.}}
}
\description{
\code{citation_metrics} adds CrossRef citation counts to the records included in the dataframe returned by \code{\link{get_records}}.
}
\seealso{
\code{\link{get_records}} for generating input dataframe.
}
