% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_records.R
\name{get_records}
\alias{get_records}
\title{Get PubMed records.}
\usage{
get_records(search_terms, pub_type = "journal article", api_key = NULL,
  date_type = "PDAT", min_date = "1966/01/01",
  max_date = format(Sys.Date(), "\%Y/\%m/\%d"))
}
\arguments{
\item{search_terms}{A character string of terms that define the scope of the PubMed database query. Boolean operators \emph{(AND, OR, NOT)} and search field tags may be used to create more complex search criteria. Commonly used search fields tags include:
\describe{
\item{TI}{Word in title}
\item{TIAB}{Word in title or abstract}
\item{MH}{Medical Subject Heading (MeSH)}
\item{AU}{Author name (e.g., Doe J)}
\item{AD}{Author institutional affiliation}
\item{TA}{Journal title (e.g., J Pain)}
}

For a full set of search fields tags: \href{https://www.ncbi.nlm.nih.gov/books/NBK3827/#_pubmedhelp_Search_Field_Descriptions_and_}{PubMed search field tags}. Note that the article publication type, date type, and date range are modified using the \code{pub_type}, \code{date_type}, \code{min_date} and \code{max_date} arguments below.}

\item{pub_type}{A character string specifying the type of publication the search must return. The default value is \emph{journal article}. For more information: \href{https://www.ncbi.nlm.nih.gov/books/NBK3827/table/pubmedhelp.T.publication_types/?report=objectonly}{PubMed article types}.}

\item{api_key}{An API character string obtained from the users NCBI account. The key is not essential, but it specifying a key gives substantially faster record query rates.}

\item{date_type}{A character string specifying the publication date type that is being specified in the search. Available values are:
\describe{
\item{PDAT}{Date the article was published \emph{(default)}.}
\item{MDAT}{Date the PubMed entry was modified.}
\item{EDAT}{Date the entry was added to PubMed.}}}

\item{min_date}{A character string in the format \emph{'YYYY/MM/DD', 'YYYY/MM' or 'YYYY'} specifying the starting date of the search. The default value is \emph{1966/01/01'}.}

\item{max_date}{A character string in the format \emph{'YYYY/MM/DD', 'YYYY/MM' or 'YYYY'} specifying the end date of the search. The default value is \code{Sys.Date()}.}
}
\value{
A long-format dataframe (each author of an article on a new row) with the following columns:
\describe{
\item{surname}{Character string specifying author surname.}
\item{initials}{Character string specifying author initials.}
\item{title}{Character string specifying title of an article.}
\item{journal}{Character string specifying the journal title (abbreviated title name).}
\item{status}{Character string specifying publication status of an article (e.g., ahead of print.}
\item{volume}{Character string specifying the journal volume.}
\item{pages}{Character string specifying the page numbers of in print articles.}
\item{year_published}{Numeric specifying the year an article came out in print.}
\item{year_online}{Numeric specifying the year an article was added to the PubMed database.}
\item{pmid}{Character string specifying the PubMed ID of an article.}
\item{doi}{Character string specifying the DOI of an article.}
\item{abstract}{Character string containing the full abstract of an article}}
}
\description{
A wrapper around \code{\link[rentrez]{entrez_fetch}}, \code{get_records} fetches \href{http://www.ncbi.nlm.nih.gov/pubmed}{PubMed} records and stores the records in a tidy dataframe suitable for processing using tools from the \emph{tidyverse}.
}
\seealso{
\code{\link[rentrez]{entrez_search}} and \code{\link[rentrez]{entrez_fetch}}

Other related functions: \code{\link{count_records}}
}
\concept{related functions}
