% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treede.R
\name{treede}
\alias{treede}
\title{PubMed xml record for Prof RD Treede.}
\source{
\url{'https://eutils.ncbi.nlm.nih.gov/entrez/eutils/esearch.fcgi?db=pubmed&id=27941497,27902566,27893485,27829508,27820160,27731639,27695361,27643836,27372299,27322440,27306631,27282592,27115670,27093432,27075429,26987721,26956043,26935254,26925007,26859819,26826097,26817644,26630440,26580679,26561600,26492533,26454224,26397930,26379077,26334010,26176894,26075963,26001168,25943423,25906795,25854794,25851512,25790450,25403802,25330039,25195796,24974768,24971044,24939241,24865794,24614828,24525274,24495935,26401296,24333949,24076047,23999056,23860634,23800791,23769719,23748119,23742795,23707287,23707286,23707285,23678019,23657118,23590332,23507585,23444968,23419598,23360414,23279302,23040698,23037991,22916273,22592181,22535540,22480770,22310319,22208803,22197694,22182852,22178393,22154921,22032740,22032306,21992218,21938604,21887117,21841049,21796057,21642846,21635191,21593389,21487289,21468319,21440369,21330175,21293898,21237569,21145371,20965658,20886070,20861430,20851519,20739597,20627413,20570558,20298428,20194147,20194146,20200655,20194142,20082204,19864066,19801050,19801048,19767146,19703571,19695306,19590390,19439674,19357343,19345493,19289256,19269743,19244350,19184121,19133877,19019713,18926637,18786858,18604567,18602755,18565660,18565003,18486546,18423989,18302701,18003941,17924294,17920770,17898403,17581853,17570589,17533117,17504645,17449306,17434803,17215503,17123732,17084865,17046675,17021023,16973292,16899640,16889902,16754839,16716518,16697110,16337817,16291301,16269136,15983783,15979027,15824343,15808988,15629873,15275800,15275782,15261863,15135912,15127169,14985260,14749441,14706490,14678844,14625457,14602841,14586594,14527605,12880820,12804667,12799965,12783950,12237192,11932070,11849764,11522578,11323151,11310627,11243472,11060130,10924804,10817609,10727924,10688052,10601424,10581220,10467918,10399864,10389177,10348320,10068155,9872431,9831453,9789583,9744923,9729360,9585326,9520240,9344076,9295212,9280161,9003568,8897487,8931593,8858495,8681865,7562615,7562614,11221788,7632416,7776255,7621362,8308146,7688283,8393170,1455981,1436631,1606484,1574584,1664508,1802341,1884168,1848966,1763252,2283538,2393105,2234499,2460316,3225589,3610490,3622675,3752481,4011284,3860887,6330009,6086371,6669318,6417623,7208079&retmode=xml&retmax=248'}
}
\usage{
treede(file_name = "treede.xml")
}
\arguments{
\item{file_name}{character string specifying the name of the demonstration xml record \emph{(treede.xml)}.}
}
\value{
An xml file with 248 PubMed records.
}
\description{
Returns an xml dataset containing journal articles (with abstracts) for Prof Rolf-Detlef Treede [President of the International Association for the Study of Pain (\href{https://www.iasp-pain.org}{IASP}): 2014-2016] for the period 1 January 1980 to 1 January 2017. The articles were retrieved from the PubMed database through the E-utilities public API of the NCBI Entrez system using \code{\link[xml2]{read_xml}} and the query string listed under \strong{'Source'} below.
}
\examples{
# Get demonstration dataset (downloaded PubMed xml query)
foo <- treede()
# Get first-author primary affiliation from dataset
parse_affiliations(foo)
# Get bibliographic information from dataset
parse_bibliographics(foo)

}

