% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_citations.R
\name{get_citations}
\alias{get_citations}
\title{Fetch CrossRef bibliometrics for PubMed records.}
\usage{
get_citations(df)
}
\arguments{
\item{df}{The dataframe object returned by \code{\link{get_records}}.}
}
\value{
Returns a dataframe with all the columns of the input dataframe, plus:
\describe{
\item{crossref_citations}{Numeric value specifying the CrossRef citation count for an article.}}
}
\description{
\code{get_citations} adds CrossRef citation counts to the records included in the dataframe returned by \code{\link{get_records}}.
}
\seealso{
\code{\link{get_records}} for generating input dataframe.
}
